--- src/main/java/com/example/springboard/dto/reqeust/article/ModiyRequest.java
@@ -1,4 +1,0 @@
-package com.example.springboard.dto.reqeust.article;
-
-public class ModiyRequest {
-}
