--- src/main/java/com/example/springboard/dto/reqeust/auth/LoginRequest.java
@@ -1,4 +1,0 @@
-package com.example.springboard.dto.reqeust.auth;
-
-public class LoginRequest {
-}
