--- src/main/java/com/example/springboard/dto/reqeust/board/ListRequest.java
@@ -1,4 +1,0 @@
-package com.example.springboard.dto.reqeust.board;
-
-public class ListRequest {
-}
