--- src/main/java/com/example/springboard/dto/reqeust/board/ManageCreateRequest.java
@@ -1,4 +1,0 @@
-package com.example.springboard.dto.reqeust.board;
-
-public class ManageCreateRequest {
-}
