--- src/main/java/com/example/springboard/dto/reqeust/board/ManageModifyScopeRequest.java
@@ -1,4 +1,0 @@
-package com.example.springboard.dto.reqeust.board;
-
-public class ManageModifyScopeRequest {
-}
