--- src/main/java/com/example/springboard/dto/reqeust/comment/AuthIsMyCommentRequest.java
@@ -1,4 +1,0 @@
-package com.example.springboard.dto.reqeust.comment;
-
-public class AuthIsMyCommentRequest {
-}
