--- src/main/java/com/example/springboard/dto/reqeust/comment/ModifyRequest.java
@@ -1,4 +1,0 @@
-package com.example.springboard.dto.reqeust.comment;
-
-public class ModifyRequest {
-}
