--- src/main/java/com/example/springboard/dto/reqeust/user/JoinCheckRequest.java
@@ -1,4 +1,0 @@
-package com.example.springboard.dto.reqeust.user;
-
-public class JoinCheckRequest {
-}
