--- src/main/java/com/example/springboard/dto/reqeust/user/ManageFindIdRequest.java
@@ -1,4 +1,0 @@
-package com.example.springboard.dto.reqeust.user;
-
-public class ManageFindIdRequest {
-}
