+++ src/main/java/com/example/springboard/dto/request/article/CreateRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.article;
+
+public class CreateRequest {
+}
