+++ src/main/java/com/example/springboard/dto/request/article/ModiyRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.article;
+
+public class ModiyRequest {
+}
