+++ src/main/java/com/example/springboard/dto/request/auth/LoginRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.auth;
+
+public class LoginRequest {
+}
