+++ src/main/java/com/example/springboard/dto/request/board/ArticleListRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.board;
+
+public class ArticleListRequest {
+}
