+++ src/main/java/com/example/springboard/dto/request/board/ManageCreateRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.board;
+
+public class ManageCreateRequest {
+}
