+++ src/main/java/com/example/springboard/dto/request/comment/CreateRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.comment;
+
+public class CreateRequest {
+}
