+++ src/main/java/com/example/springboard/dto/request/comment/ModifyRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.comment;
+
+public class ModifyRequest {
+}
