+++ src/main/java/com/example/springboard/dto/request/user/JoinCheckRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.user;
+
+public class JoinCheckRequest {
+}
