+++ src/main/java/com/example/springboard/dto/request/user/ManageDeleteAccountRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.user;
+
+public class ManageDeleteAccountRequest {
+}
