+++ src/main/java/com/example/springboard/dto/request/user/ManageFindIdRequest.java
@@ -1,0 +1,4 @@
+package com.example.springboard.dto.request.user;
+
+public class ManageFindIdRequest {
+}
